% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsViolin.R
\docType{package}
\name{dotsViolin}
\alias{dotsViolin}
\title{Integrates tables and plots}
\description{
A series of functions to get modes/peaks from
discrete and continuous variables and integrate them
as tables inside plots
cite as in: citation("dotsViolin")
}
