% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabaceae_genome_size.R
\docType{data}
\name{fabaceae_clade_1Cx_df}
\alias{fabaceae_clade_1Cx_df}
\title{Genome sizes for fabaceae}
\format{
data.frame with columns:
\describe{
  \item{name}{OTU, species}
  \item{clade}{main fabaceae clade}
  \item{Cx}{genome size, Cx}
}
}
\usage{
fabaceae_clade_1Cx_df
}
\description{
fabaceae_clade_1Cx_df: parsed Cx sizes for fabaceae
}
\seealso{
\code{\link{get_peaks_counts_continuous}}
}
\keyword{datasets}
