% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabaceae_clade_chrom.R
\docType{data}
\name{fabaceae_clade_n_df}
\alias{fabaceae_clade_n_df}
\title{chromosomal counts for fabaceae}
\format{
data.frame with columns:
\describe{
  \item{tip.label}{OTU, species}
  \item{clade}{main fabaceae clade}
  \item{parsed_n}{chromosome number, n}
}
}
\usage{
fabaceae_clade_n_df
}
\description{
fabaceae_clade_n_df: parsed n counts for fabaceae
}
\seealso{
\code{\link{get_modes_counts}}
}
\keyword{datasets}
