% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.peaks.R
\name{get.peaks}
\alias{get.peaks}
\title{Get peaks of a continuous variable}
\usage{
get.peaks(x, bw, signifi, nsmall, ranks = 3)
}
\arguments{
\item{x}{dataframe}

\item{bw}{bandwidth}

\item{signifi}{criteria to bin the data in number of digits}

\item{nsmall}{criteria to approximate (round) data}

\item{ranks}{numeric how many ranks to consider}
}
\value{
data.frame
}
\description{
This function allows you to get peaks for a continuous variable.
Based on the kernel density function
}
\keyword{modes}
