% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modes_counts.R
\name{get_modes_counts}
\alias{get_modes_counts}
\title{get modes, handle ties, ignore less frequent values}
\usage{
get_modes_counts(data, grouping_col, col2, mode_number = 3)
}
\arguments{
\item{data}{data.frame}

\item{grouping_col}{string split by this column}

\item{col2}{string numerical data column}

\item{mode_number}{numeric number of modes to retrieve}
}
\value{
data.frame with modes and counts per group
}
\description{
This function comes from an answer for a question in stackoverflow
https://stackoverflow.com/questions/42698465/obtaining-3-most-common-elements-of-groups-concatenating-ties-and-ignoring-les
}
\examples{
get_modes_counts(fabaceae_clade_n_df, "clade", "parsed_n")
}
