% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_counts_continuous.R
\name{get_peaks_counts_continuous}
\alias{get_peaks_counts_continuous}
\title{Peaks of a continuous variable in a dataframe format}
\usage{
get_peaks_counts_continuous(
  origtable,
  grouping_col,
  columnname,
  peak_number,
  adjust1,
  signifi,
  nsmall
)
}
\arguments{
\item{origtable}{dataframe}

\item{grouping_col}{column with categories - character}

\item{columnname}{column with numerical data}

\item{peak_number}{number of peaks to get, see get.peaks}

\item{adjust1}{bandwith adjust parameter}

\item{signifi}{see get.peaks function}

\item{nsmall}{see get.peaks function}
}
\value{
data.frame
}
\description{
This function allows you to get peaks and summary counts per group for a continuos variable in a dataframe format.
Handles ties; least frequent is ignored, except if it is the only
one, depends on \code{get.peaks} function
}
\examples{
get_peaks_counts_continuous(fabaceae_clade_1Cx_df, "clade", "Cx", 2, 0.25, 1, 2)
}
\keyword{peaks}
