% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotviolin.R
\name{plot_dotviolin}
\alias{plot_dotviolin}
\title{Makes a dot-plot and violin-plot}
\usage{
plot_dotviolin(
  dataset,
  par,
  groupcol,
  vary,
  labelx,
  maxx,
  adjust,
  binwidth,
  fill_group = "fill_group",
  font = "mono",
  dots = TRUE,
  violin = TRUE
)
}
\arguments{
\item{dataset}{dataframe with columns to be merged into 1}

\item{par}{dot size}

\item{groupcol}{categories to group}

\item{vary}{numeric variable}

\item{labelx}{x axis label}

\item{maxx}{x axis maximum value}

\item{adjust}{geom_violin adjust parameter}

\item{binwidth}{geom_dotplot binwidth parameter}

\item{fill_group}{2nd category with 2 options as a fill aes argument for geom_dotplot}

\item{font}{font family}

\item{dots}{boolean include dot plot}

\item{violin}{boolean include violin plot}
}
\value{
ggplot
}
\description{
This function makes a dot-plot and violin-plot, internal function
}
\keyword{dot_plots}
\keyword{violin}
